Added feature options
Added `anywhere_on_line` option that will show the virtual text if the cursor is on a line instead of just on the cursor.

Added `show_same_line_opening` option that allows you to hide the virtual text if the opening paren is on the same line as the close.